--- packages/app/components/activity/index.tsx
@@ -1,1 +1,0 @@
-export { Activity } from "./activity";
