--- packages/app/components/activity/types/comment.tsx
@@ -1,91 +1,0 @@
-import { TextLink } from "app/navigation/link";
-
-function Comment({ act }) {
-  const { nfts } = act;
-  const count = nfts?.length;
-
-  return (
-    <>
-      <>
-        {count === 1 && (
-          <>
-            commented on{" "}
-            <TextLink
-              tw="text-sm font-bold text-black dark:text-white"
-              href={`/nft/${nfts[0].nft_id}`}
-            >
-              {nfts[0].token_name}
-            </TextLink>
-            .
-          </>
-        )}
-        {count === 2 && (
-          <>
-            commented on{" "}
-            <TextLink
-              tw="text-sm font-bold text-black dark:text-white"
-              href={`/nft/${nfts[0].nft_id}`}
-            >
-              {nfts[0].token_name}
-            </TextLink>{" "}
-            and{" "}
-            <TextLink
-              tw="text-sm font-bold text-black dark:text-white"
-              href={`/nft/${nfts[1].nft_id}`}
-            >
-              {nfts[1].token_name}
-            </TextLink>
-            .
-          </>
-        )}
-        {count === 3 && (
-          <>
-            commented on{" "}
-            <TextLink
-              tw="text-sm font-bold text-black dark:text-white"
-              href={`/nft/${nfts[0].nft_id}`}
-            >
-              {nfts[0].token_name}
-            </TextLink>
-            ,{" "}
-            <TextLink
-              tw="text-sm font-bold text-black dark:text-white"
-              href={`/nft/${nfts[1].nft_id}`}
-            >
-              {nfts[1].token_name}
-            </TextLink>{" "}
-            and{" "}
-            <TextLink
-              tw="text-sm font-bold text-black dark:text-white"
-              href={`/nft/${nfts[2].nft_id}`}
-            >
-              {nfts[2].token_name}
-            </TextLink>
-            .
-          </>
-        )}
-        {count > 3 && (
-          <>
-            commented on{" "}
-            <TextLink
-              tw="text-sm font-bold text-black dark:text-white"
-              href={`/nft/${nfts[0].nft_id}`}
-            >
-              {nfts[0].token_name}
-            </TextLink>
-            ,{" "}
-            <TextLink
-              tw="text-sm font-bold text-black dark:text-white"
-              href={`/nft/${nfts[1].nft_id}`}
-            >
-              {nfts[1].token_name}
-            </TextLink>{" "}
-            and {count - 2} others.
-          </>
-        )}
-      </>
-    </>
-  );
-}
-
-export { Comment };
