--- packages/app/components/activity/types/index.ts
@@ -1,7 +1,0 @@
-export { Buy } from "./buy";
-export { Comment } from "./comment";
-export { Create } from "./create";
-export { Follow } from "./follow";
-export { Like } from "./like";
-export { Sell } from "./sell";
-export { Transfer } from "./transfer";
