--- packages/app/components/camera/index.tsx
@@ -1,1 +1,0 @@
-export { Camera } from "./vision-camera";
