--- packages/app/components/camera/index.web.tsx
@@ -1,1 +1,0 @@
-export { Camera } from "./expo-camera";
