--- packages/app/components/camera/vision-camera.tsx
@@ -1,7 +1,0 @@
-// Multi camera on Android not yet supported by CameraX
-// "Thanks for the request. Currently CameraX does not support the multi camera API but as more device adopt them, we will enable support at the appropriate time. Thanks."
-// https://issuetracker.google.com/issues?q=componentid:618491%20status:open&pli=1
-
-export function Camera() {
-  return null;
-}
