--- packages/app/components/nft-activity/nft-activity.constants.tsx
@@ -1,11 +1,0 @@
-export const BLOCKCHAIN_TYPES = {
-  POLYGONSCAN: "polygonscan",
-};
-
-export const EVENT_TYPES = {
-  CREATED: "Created",
-  SALE: "Sale",
-  TRANSFER: "Transfer",
-  LIST: "List",
-  MINTED: "Minted",
-};
