+++ packages/design-system/bottom-sheet/bottom-sheet-modal-provider.tsx
@@ -1,0 +1,3 @@
+import { BottomSheetModalProvider } from "@gorhom/bottom-sheet";
+
+export { BottomSheetModalProvider };
