+++ packages/design-system/bottom-sheet/bottom-sheet-modal-provider.web.tsx
@@ -1,0 +1,4 @@
+import { Fragment } from "react";
+
+const BottomSheetModalProvider = Fragment;
+export { BottomSheetModalProvider };
