+++ packages/design-system/bottom-sheet/bottom-sheet-scroll-view.tsx
@@ -1,0 +1,3 @@
+import { BottomSheetScrollView } from "@gorhom/bottom-sheet";
+
+export { BottomSheetScrollView };
