+++ packages/design-system/bottom-sheet/bottom-sheet-scroll-view.web.tsx
@@ -1,0 +1,4 @@
+import { ScrollView } from "react-native";
+
+const BottomSheetModalProvider = ScrollView;
+export { BottomSheetModalProvider };
