+++ packages/design-system/bottom-sheet/bottom-sheet.web.tsx
@@ -1,0 +1,5 @@
+import { Fragment } from "react";
+
+export const BottomSheet = Fragment;
+
+export const BottomSheetTextInput = Fragment;
